chore(team editor list): use link component for team name
